+++ main.py
@@ -1,6 +1,8 @@
 import flask
-import pages.home
+from src.views import views
 
-app = flask.Flask(__name__, template_folder="templates")
+app = flask.Flask(__name__)
+app.register_blueprint(views)
 
-app.register_blueprint(pages.home.home)
+if __name__ == "__main__":
+    app.run(debug=True)