+++ pages/home/__init__.py
@@ -1,5 +1,0 @@
-import flask
-from ._view import *
-
-home = flask.Blueprint("home", __name__)
-home.register_blueprint(view)